requirements.txt
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
plotly>=5.3.0
pydicom>=2.3.0
nibabel>=4.0.0
tqdm>=4.65.0
pyyaml>=6.0
openai>=0.27.0
